requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
redis[hiredis]==5.0.7
pyjwt==2.8.0
pyyaml==6.0.1
structlog==24.2.0
prometheus-client==0.20.0
python-dotenv==1.0.1